// src/types/table.d.ts
export interface TableRowData {
  runningAge: string | number;
  year: string | number;
  month: string | number;
  N: string | number;
  Nx: string | number;
  yearly_expenses: string | number;
  monthly_expenses: string | number;
  monthly_income: string | number;
  income_Nx: string | number;
  expected_income: string | number;
  income_invest: string | number;
  actual_income?: string | number;
  actual_expenses?: string | number;
  actual_return?: string | number;
}